+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # stripe-angular - Change Log
 All notable changes to this project will be documented here.
 
+## [1.5.1] - (2021-01-04)
+- Create Card Payment Method
+- Stripe data typings instead of created ones
+- Demo area has much more testing functionality
+
 ## [1.4.3] - (2020-11-19)
 - more demo support
 - fixes